Prevent runtime restart for crashed clients to framework sensorservice.
This only happens when the client dies, in which case the event queue
reference will be dropped on the client side, so the call to handleEvent
won't even be fired, except for a race condition.

The race condition is, if handleEvent is already fired,
then client dies and the event queue is going to be destroyed and the
fd is going to be removed, but before that the call onEvent fails.

Test: clients of SensorManager that uses event queue no longer
      crashes.
Bug: 35219747

Change-Id: Ibdf062528a219fae65ee6263ed742229ec3cd0d6