Fix a anr bug caused by sendFinishedSignal logical error
Because of eliminate multiple benign overflow conditions see:
https://android-review.googlesource.com/#/c/172237/
Changed the do while loop resulting in a logical difference.

When the chainIndex-- to 0, the loop is not run.
It is the most important cycle, it will push head seq to mSeqChains.
If not run, will lead to a batch of seq can not corrected finish.
Eventually leading to the occurrence of anr.

Signed-off-by: gaoshang <gaoshang@xiaomi.com>
Test: Input dispatche process
Bug: 38366215
Change-Id: I87e609dfcb00ac7b8e82c6de789df094e7c25efd