libvrflinger: Inject USAGE_HW_COMPOSER usage into DirectSurface buffers.
DirectSurfaces are used by the VR display manager to do direct rendering
in VR mode. The buffers backing a DirectSurface must include the HWC
usage flag, but this flag is not exposed through AHardwareBuffer, for
good reason. Inject this necessary usage bit on the framework side to
enable proper operation.

Bug: None
Test: Build; run.
Change-Id: I46e6b97e41763aaa7870fa64b09ea3a5a81990c1