Do not set the trace clock in atrace.rc.
This reverts only the atrace.rc change from commit ea826794184be368f65f099ceac4fc31234700f6.
"Set the atrace clock to the best available value: boot, mono, or global."

This change caused a 100+ms regression in boot time. We also confirm that
the clock is set to the correct value when we call atrace, so we will
still see correct behavior after reverting this change.

Bug: 38504158
Test: Manually tested that the trace clock is updated when atrace is
called.
Change-Id: Ifa302694964a4f9fc54db16e7125d5279067bbfd