Print "test skipped" for wide-color tests
Currently if EGL_test is run on non-wide color device the
test will print a status of "PASS". Which is correct, but
not as useful as it could be. Added a diagnostic message
in that case indicating that the device is not wide-color
so that I can tell which path was taken in the tests.
Bug: 63077212
Test: adb shell /data/nativetest/EGL_test/EGL_test

Change-Id: Ic65316893adb4273f3652a3613440ca02001079c