Update LayerVector's order after modifying relative layer.
If the z value for a layer is updated due to a relative layer change,
the order of layers in the SortedVector was not updated and could cause
lookup failures in the binary search. This updates the vector by
removing and then re-adding the layer to ensure it's resorted.

Also added the eRelativeLayerChanged condition for SurfaceFlinger_hwc1.

Test: Added multiple layers that updated relative layer. The LayerVector's
      order was updated correctly and the layers remain in the correct
      order.

Change-Id: If5403a0b626f34db41bf7a98fe42830ed272d16e