Properly applies the selected saturation boost (vivid mode)
The saturation boost setting is read by SurfaceFlinger as a
persistent system property. Unfortunately, persistent props
are only available after Vold is up and /data is decrypted,
which may happen before or after SF attempts to read the
property.

This CL moves the propery lookup to the end of the boot
animation. This solves two issues:
- The saturation boost will not be applied to the boot animation
- The vivid colors user setting is now reliably applied

Bug: 63823274
Test: Manual

Change-Id: Icb8e30c799c30cf674f0fc0bab0369f4c99367ed